+++ src/components/calendar/upcomingevents/googleeventprops.ts
@@ -1,0 +1,13 @@
+export type GoogleEventProps = {
+  start: {
+    dateTime?: string;
+    date?: string;
+  };
+  end: {
+    dateTime?: string;
+    date?: string;
+  };
+  location?: string;
+  description?: string;
+  summary: string;
+};
